+++ src/components/shared/project-card.tsx
@@ -52,7 +52,10 @@ const ProjectCard = ({ projeto }: ProjectCardProps) => {
           <div className="flex items-center space-x-2 text-sm text-muted-foreground mt-4 pt-2">
             <Avatar className="h-6 w-6">
               <AvatarImage
-                src={projeto.autor.urlFotoPerfil || getDefaultAvatarUrl(projeto.autor.id)}
+                src={
+                  projeto.autor.urlFotoPerfil ||
+                  getDefaultAvatarUrl(projeto.autor.id, projeto.autor.nome)
+                }
                 alt={projeto.autor.nome}
               />
               <AvatarFallback>{projeto.autor.nome[0]}</AvatarFallback>